+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### [2026-09-12]
+
+#### Added
+
+- Menambahkan Unit Test dan Integration Test
+
 ### [2026-09-11]
 
 #### Added